+++ moabb/tests/test_datasets.py
@@ -9,7 +9,7 @@ import pytest
 
 import moabb.datasets as db
 import moabb.datasets.compound_dataset as db_compound
-from moabb.datasets import BNCI2014_001, Cattan2019_VR, Shin2017A, Shin2017B
+from moabb.datasets import BNCI2014_001, Cattan2019_VR, Kojima2024A, Shin2017A, Shin2017B
 from moabb.datasets.base import (
     BaseDataset,
     LocalBIDSDataset,
@@ -593,3 +593,36 @@ class TestBIDSDataset:
             for session_data in subject_data.values():
                 assert session_data.keys() == {"0"}
                 assert isinstance(session_data["0"], mne.io.BaseRaw)
+
+
+class TestKojima2024A:
+    def test_convert_subject_to_subject_id(self):
+        ds = Kojima2024A()
+        assert ds.convert_subject_to_subject_id(1) == "A"
+        assert ds.convert_subject_to_subject_id(3) == "C"
+        assert ds.convert_subject_to_subject_id(list(range(1, 12))) == [
+            "A",
+            "B",
+            "C",
+            "D",
+            "E",
+            "F",
+            "G",
+            "H",
+            "I",
+            "J",
+            "K",
+        ]
+        with pytest.raises(TypeError):
+            ds.convert_subject_to_subject_id(1.5)
+
+    def test_data_shape(self):
+        ds = Kojima2024A()
+        paradigm = P300()
+        X, labels, meta = paradigm.get_data(dataset=ds, subjects=[1])
+
+        # number of channels
+        assert X.shape[1] == 64
+
+        # number of samples
+        assert X.shape[0] == len(labels)